# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Glory Sports API - Cricket sports ecosystem for India
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List user's bookings
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: turfId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a turf slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # MATCHES
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List matches
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: teamId
          in: query
          schema:
            type: integer
        - name: playerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Matches list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a new match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{matchId}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get match details
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{matchId}/score:
    patch:
      operationId: updateMatchScore
      tags: [matches]
      summary: Update match score
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchScoreInput"
      responses:
        "200":
          description: Score updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{matchId}/verify:
    post:
      operationId: verifyMatch
      tags: [matches]
      summary: Captain verifies match result
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchVerifyInput"
      responses:
        "200":
          description: Match verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{matchId}/confirm-participation:
    post:
      operationId: confirmMatchParticipation
      tags: [matches]
      summary: Player confirms they played in a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Participation confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"

  # STATS
...
components:
  schemas:
...
    Booking:
      type: object
      required: [id, turfId, userId, date, startTime, endTime, status, createdAt]
      properties:
        id:
          type: integer
        turfId:
          type: integer
        turfName:
          type: string
        userId:
          type: integer
        userName:
          type: string
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        totalAmount:
          type: number
        status:
          type: string
          enum: [confirmed, cancelled, completed]
        matchId:
          type: ["integer", "null"]
        createdAt:
          type: string
    BookingInput:
      type: object
      required: [turfId, date, startTime, endTime]
      properties:
        turfId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string

    # MATCHES
    Match:
      type: object
      required: [id, turfId, teamAId, teamBId, status, createdAt]
      properties:
        id:
          type: integer
        turfId:
          type: integer
        turfName:
          type: string
        teamAId:
          type: integer
        teamAName:
          type: string
        teamBId:
          type: integer
        teamBName:
          type: string
        overs:
          type: integer
        matchType:
          type: string
        matchDate:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, verified, disputed, flagged]
        teamAScore:
          type: ["integer", "null"]
        teamAWickets:
          type: ["integer", "null"]
        teamBScore:
          type: ["integer", "null"]
        teamBWickets:
          type: ["integer", "null"]
        winnerTeamId:
          type: ["integer", "null"]
        mvpPlayerId:
          type: ["integer", "null"]
        mvpPlayerName:
          type: ["string", "null"]
        teamACaptainVerified:
          type: boolean
        teamBCaptainVerified:
          type: boolean
        isSuspicious:
          type: boolean
        bookingId:
          type: ["integer", "null"]
        createdAt:
          type: string
    MatchInput:
      type: object
      required: [turfId, teamAId, teamBId, overs, matchDate]
      properties:
        turfId:
          type: integer
        teamAId:
          type: integer
        teamBId:
          type: integer
        overs:
          type: integer
        matchType:
          type: string
        matchDate:
          type: string
        bookingId:
          type: integer
    MatchScoreInput:
      type: object
      properties:
        teamAScore:
          type: integer
        teamAWickets:
          type: integer
        teamBScore:
          type: integer
        teamBWickets:
          type: integer
        winnerTeamId:
          type: integer
        mvpPlayerId:
          type: integer
        status:
          type: string
    MatchVerifyInput:
      type: object
      required: [verified]
      properties:
        verified:
          type: boolean
        captainSide:
          type: string
          enum: [A, B]
...