# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Glory Sports API - Cricket sports ecosystem for India
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  turfId: number;
  turfName?: string;
  userId: number;
  userName?: string;
  date: string;
  startTime: string;
  endTime: string;
  totalAmount?: number;
  status: BookingStatus;
  /** @nullable */
  matchId?: number | null;
  createdAt: string;
}