# lib/api-zod/src/generated/types/matchInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Glory Sports API - Cricket sports ecosystem for India
 * OpenAPI spec version: 0.1.0
 */

export interface MatchInput {
  turfId: number;
  teamAId: number;
  teamBId: number;
  overs: number;
  matchType?: string;
  matchDate: string;
  bookingId?: number;
}